requirements.txt
numpy
scipy
pandas
matplotlib
altair
metaflow
python-dotenv
tqdm
filelock
typer
sh
jupyter
scikit-learn==0.23.2
transformers
sentence-transformers==1.2.0
networkx
gensim
bokeh
umap-learn
nlpaug
nltk
pigeonXT-jupyter
boto3
fsspec
s3fs
geopandas
rtree
urllib3
shapely